Module sheave_client::handlers

Modules

middlewares

Structs

ConnectionError

An error that handling connect command has been failed.

PublicationError

An error that handling publish command has been failed.

RtmpHandler

The default RTMP handler.

Functions

connection_error

A utility function of constructing an ConnectionError error.

publication_error

A utility function of constructing an PublicationError error.